requirements.txt
os
openai
dotenv
discord.py
random
discord.ext
asyncio
epicstore_api
datetime
